apps/voice-agent/requirements.txt
# Voice Agent Service Dependencies
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0

# LLM - Ollama client
ollama>=0.1.7

# Speech-to-Text - Whisper (optional, install separately if needed)
# faster-whisper>=1.0.0
# Note: On Windows, you may need Rust toolchain. Install via: pip install faster-whisper
# Or use pre-built wheels if available for your Python version

# MCP client
httpx>=0.25.0
aiohttp>=3.9.0

# Audio processing
numpy>=1.26.0
pydub>=0.25.0

# Utilities
pydantic>=2.5.0

